[poisson] Add support to use a custom node-node connectivity instead of creating edges.
At the moment, custom connectivity is optional and is only used if we
add the option '<create-mesh>' to 'false' in the case file.

Because Edges are only used to know about the neighbours of a node, it is
bette to keep the information in a custom connectivity. It will reduce
both init time and memory usage.

This need at least Arcane version 3.14.4.