[femutils] Fix a strange SEGV occuring in some configuration.
With some configuration, it seems 'dof_family' in the ENUMERATE_ is null and
so there is a SEGV when using it. If we do some modification in the code
(like printing the value of 'dof_family'), the bug disappears.
It may to be a compiler bug (it crashes with GCC 13 or 14 but works with
several versions of clang).